Separate models into their own files; organize extension functions
