--- app/src/main/java/website/grahamearley/placefinder/Models.kt
@@ -1,219 +1,0 @@
-package website.grahamearley.placefinder
-
-import android.annotation.SuppressLint
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-/**
- * Data models generated by JsonToKotlinClass: https://github.com/wuseal/JsonToKotlinClass
- */
-
-data class FoursquareResponse(
-		val meta: Meta,
-		val response: Response
-)
-
-data class Meta(
-		val code: Int,
-		val requestId: String
-)
-
-data class Response(
-		val groups: List<VenueGroup>? = null,
-        val tips: Tips? = null,
-        val photos: Photos? = null
-)
-
-data class Tips(
-        val count: Int? = null,
-        val items: List<Tip>? = null
-)
-
-data class VenueGroup(
-		val type: String? = null,
-		val name: String? = null,
-		val items: List<VenueItem>? = null
-)
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class VenueItem (
-		val reasons: Reasons? = null,
-		val venue: Venue? = null,
-		val tips: List<Tip>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Reasons(
-		val count: Int? = null,
-		val items: List<ReasonItem>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class ReasonItem(
-		val summary: String? = null,
-		val type: String? = null,
-		val reasonName: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Venue(
-		val id: String? = null,
-		val name: String? = null,
-		val contact: Contact? = null,
-		val location: Location? = null,
-		val categories: List<Category?>? = null,
-		val url: String? = null,
-		val price: Price? = null,
-		val rating: Double? = null,
-		val ratingColor: String? = null,
-		val ratingSignals: Int? = null,
-		val menu: Menu? = null,
-		val hours: Hours? = null,
-		val photos: Photos? = null,
-		val featuredPhotos: FeaturedPhotos? = null) : Parcelable {
-
-    fun getFirstFeaturedPhotoOrNull() = featuredPhotos?.items?.firstOrNull()
-
-    fun getPhotoUrlsOrNull() = photos?.groups?.flatMap { it.items.orEmpty() }?.map(PhotoItem::getUrl)
-
-    fun getFirstCategoryOrNull() = categories?.firstOrNull()?.name
-
-    fun getStreetAddressOrNull() = location?.formattedAddress?.firstOrNull()
-
-}
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Price(
-		val tier: Int? = null,
-		val message: String? = null,
-		val currency: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Contact(
-		val phone: String? = null,
-		val formattedPhone: String? = null,
-		val twitter: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Category(
-		val id: String? = null,
-		val name: String? = null,
-		val pluralName: String? = null,
-		val shortName: String? = null,
-		val icon: Icon? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Icon(
-		val prefix: String? = null,
-		val suffix: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Hours(
-		val status: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class FeaturedPhotos(
-		val count: Int? = null,
-		val items: List<PhotoItem?>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Location(
-		val address: String? = null,
-		val crossStreet: String? = null,
-		val lat: Double? = null,
-		val lng: Double? = null,
-		val postalCode: String? = null,
-		val cc: String? = null,
-		val city: String? = null,
-		val state: String? = null,
-		val country: String? = null,
-		val formattedAddress: List<String?>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Menu(
-		val type: String? = null,
-		val label: String? = null,
-		val anchor: String? = null,
-		val url: String? = null,
-		val mobileUrl: String? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Photos(
-		val count: Int? = null,
-		val groups: List<PhotoGroup>? = null,
-        val items: List<PhotoItem>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class PhotoGroup(
-		val type: String? = null,
-		val name: String? = null,
-		val count: Int? = null,
-		val items: List<PhotoItem>? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class PhotoItem(
-		val id: String? = null,
-		val createdAt: Int? = null,
-		val prefix: String? = null,
-		val suffix: String? = null,
-		val width: Int? = null,
-		val height: Int? = null,
-		val user: User? = null,
-		val visibility: String? = null): Parcelable {
-
-    fun getUrl(): String {
-        // cap300 => photo size with width or height of 300 (whichever is larger)
-        return prefix + "cap300" + suffix
-    }
-
-}
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class Tip(
-		val id: String? = null,
-		val createdAt: Long? = null,
-		val text: String? = null,
-		val type: String? = null,
-		val url: String? = null,
-		val canonicalUrl: String? = null,
-		val agreeCount: Int? = null,
-		val disagreeCount: Int? = null,
-		val user: User? = null
-): Parcelable
-
-@SuppressLint("ParcelCreator")
-@Parcelize
-data class User(
-		val id: String,
-		val firstName: String? = null,
-        val lastName: String? = null,
-		val gender: String? = null,
-		val photo: PhotoItem? = null,
-		val type: String? = null
-): Parcelable
