+++ app/src/main/java/website/grahamearley/placefinder/extension/ContextExtensions.kt
@@ -1,0 +1,36 @@
+package website.grahamearley.placefinder.extension
+
+import android.app.Activity
+import android.app.ActivityOptions
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.support.annotation.StringRes
+import android.view.inputmethod.InputMethodManager
+import android.widget.Toast
+
+/**
+ * Extension functions for Android context classes.
+ */
+
+fun Activity.hideSoftKeyboard() {
+    val view = this.currentFocus
+
+    if (view != null) {
+        val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+        imm.hideSoftInputFromWindow(view.windowToken, 0)
+    }
+}
+
+fun Activity.startActivityWithTransitionIfPossible(intent: Intent) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+        startActivity(intent,
+                ActivityOptions.makeSceneTransitionAnimation(this).toBundle())
+    } else {
+        startActivity(intent)
+    }
+}
+
+fun Context.showToast(@StringRes stringRes: Int) {
+    Toast.makeText(this, stringRes, Toast.LENGTH_LONG).show()
+}
