+++ app/src/main/java/website/grahamearley/placefinder/extension/ViewExtensions.kt
@@ -1,32 +1,19 @@
-package website.grahamearley.placefinder
+package website.grahamearley.placefinder.extension
 
-import android.app.Activity
-import android.app.ActivityOptions
-import android.content.Context
-import android.content.Intent
-import android.os.Build
+import android.graphics.drawable.BitmapDrawable
 import android.support.annotation.DimenRes
 import android.support.annotation.StringRes
 import android.support.design.widget.Snackbar
+import android.support.v4.graphics.drawable.RoundedBitmapDrawableFactory
 import android.support.v4.view.ViewCompat
 import android.support.v7.content.res.AppCompatResources
 import android.view.View
-import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
-import android.widget.Toast
 import com.squareup.picasso.Picasso
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import android.opengl.ETC1.getHeight
-import android.opengl.ETC1.getWidth
-import android.support.v4.graphics.drawable.RoundedBitmapDrawableFactory
-import android.support.v4.graphics.drawable.RoundedBitmapDrawable
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.Bitmap
+import website.grahamearley.placefinder.R
 
 /**
- * Utility extension functions for various classes.
+ * Extension functions for views.
  */
 
 fun View.show() {
@@ -72,28 +59,6 @@ fun ImageView.loadImageRounded(url: String) {
                     this@loadImageRounded.setImageDrawable(imageDrawable)
                 }
             })
-}
-
-fun Activity.hideSoftKeyboard() {
-    val view = this.currentFocus
-
-    if (view != null) {
-        val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-        imm.hideSoftInputFromWindow(view.windowToken, 0)
-    }
-}
-
-fun Context.showToast(@StringRes stringRes: Int) {
-    Toast.makeText(this, stringRes, Toast.LENGTH_LONG).show()
-}
-
-fun Activity.startActivityWithTransitionIfPossible(intent: Intent) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-        startActivity(intent,
-                ActivityOptions.makeSceneTransitionAnimation(this).toBundle())
-    } else {
-        startActivity(intent)
-    }
 }
 
 fun View.showSnackbar(@StringRes stringRes: Int) {
