+++ app/src/main/java/website/grahamearley/placefinder/model/FoursquareResponse.kt
@@ -1,0 +1,10 @@
+package website.grahamearley.placefinder.model
+
+/**
+ * Data models generated by JsonToKotlinClass: https://github.com/wuseal/JsonToKotlinClass
+ */
+
+data class FoursquareResponse(
+        val meta: Meta,
+        val response: Response
+)
