+++ app/src/main/java/website/grahamearley/placefinder/model/Hours.kt
@@ -1,0 +1,11 @@
+package website.grahamearley.placefinder.model
+
+import android.annotation.SuppressLint
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@SuppressLint("ParcelCreator")
+@Parcelize
+data class Hours(
+		val status: String? = null
+): Parcelable
