+++ app/src/main/java/website/grahamearley/placefinder/model/Meta.kt
@@ -1,0 +1,6 @@
+package website.grahamearley.placefinder.model
+
+data class Meta(
+		val code: Int,
+		val requestId: String
+)
