+++ app/src/main/java/website/grahamearley/placefinder/model/PhotoGroup.kt
@@ -1,0 +1,14 @@
+package website.grahamearley.placefinder.model
+
+import android.annotation.SuppressLint
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@SuppressLint("ParcelCreator")
+@Parcelize
+data class PhotoGroup(
+		val type: String? = null,
+		val name: String? = null,
+		val count: Int? = null,
+		val items: List<PhotoItem>? = null
+): Parcelable
