+++ app/src/main/java/website/grahamearley/placefinder/model/Response.kt
@@ -1,0 +1,7 @@
+package website.grahamearley.placefinder.model
+
+data class Response(
+        val groups: List<VenueGroup>? = null,
+        val tips: Tips? = null,
+        val photos: Photos? = null
+)
