+++ app/src/main/java/website/grahamearley/placefinder/model/Tips.kt
@@ -1,0 +1,6 @@
+package website.grahamearley.placefinder.model
+
+data class Tips(
+        val count: Int? = null,
+        val items: List<Tip>? = null
+)
