+++ app/src/main/java/website/grahamearley/placefinder/model/VenueGroup.kt
@@ -1,0 +1,7 @@
+package website.grahamearley.placefinder.model
+
+data class VenueGroup(
+		val type: String? = null,
+		val name: String? = null,
+		val items: List<VenueItem>? = null
+)
