+++ app/src/main/java/website/grahamearley/placefinder/ui/detail/PlaceDetailActivity.kt
@@ -10,6 +10,9 @@ import android.support.v7.widget.LinearLayoutManager
 import android.view.MenuItem
 import kotlinx.android.synthetic.main.activity_place_detail.*
 import website.grahamearley.placefinder.*
+import website.grahamearley.placefinder.extension.*
+import website.grahamearley.placefinder.model.Tip
+import website.grahamearley.placefinder.model.VenueItem
 import website.grahamearley.placefinder.ui.detail.contract.PlaceDetailPresenterContract
 import website.grahamearley.placefinder.ui.detail.contract.PlaceDetailViewContract
