+++ app/src/main/java/website/grahamearley/placefinder/ui/detail/TipRecyclerAdapter.kt
@@ -4,6 +4,10 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import website.grahamearley.placefinder.*
+import website.grahamearley.placefinder.extension.hide
+import website.grahamearley.placefinder.extension.loadImageRounded
+import website.grahamearley.placefinder.extension.show
+import website.grahamearley.placefinder.model.Tip
 import java.text.SimpleDateFormat
 import java.util.*
