+++ app/src/main/java/website/grahamearley/placefinder/ui/list/PlaceListPresenter.kt
@@ -4,9 +4,8 @@ import io.reactivex.Scheduler
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
-import website.grahamearley.placefinder.FoursquareResponse
 import website.grahamearley.placefinder.R
-import website.grahamearley.placefinder.VenueItem
+import website.grahamearley.placefinder.model.VenueItem
 import website.grahamearley.placefinder.data.FoursquareInteractor
 import website.grahamearley.placefinder.data.FoursquareInteractorContract
 import website.grahamearley.placefinder.ui.list.contract.PlaceListPresenterContract
